+++ Provider/src/main/java/org/example/provide/Eur.java
@@ -7,7 +7,7 @@ import org.example.convert.CurrencyFullName;
 public class Eur implements Converter {
 
     @Override
-    public double convertMoney(int givenAmount) {
+    public double convertMoney(float givenAmount) {
         return givenAmount / 9.76;
     }
